refactor: move ops around
It is not txn specific, so move to a more generic location.
