+++ packages/coln-store/src/txn/row_handle.rs
@@ -7,11 +7,10 @@ use std::{cell::RefCell, rc::Rc};
 use crate::{
     commit::hash::CommitHash,
     ir,
+    op::Op,
     store::error::StoreIntError,
-    table::{CellValue, PackedCell, PackedRowId, RowId, ValidationError},
+    table::{CellValue, RowId, ValidationError},
 };
-
-pub const OP_KIND_ADD: u32 = 0;
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
 pub struct TxnId(u64);
@@ -161,9 +160,7 @@ impl From<RowHandle> for TxnValue {
 
 impl From<RowId> for TxnValue {
     fn from(value: RowId) -> Self {
-        TxnValue::Id(RowHandle {
-            state: Rc::new(RefCell::new(RowHandleState::Existing(value))),
-        })
+        TxnValue::Id(RowHandle::from_existing(value))
     }
 }
 
@@ -188,19 +185,14 @@ impl From<&str> for TxnValue {
 impl From<CellValue> for TxnValue {
     fn from(value: CellValue) -> Self {
         match value {
-            CellValue::Id(id) => TxnValue::Id(RowHandle {
-                state: Rc::new(RefCell::new(RowHandleState::Existing(id))),
-            }),
+            CellValue::Id(id) => TxnValue::Id(RowHandle::from_existing(id)),
             CellValue::Int(value) => TxnValue::Int(value),
             CellValue::Str(value) => TxnValue::Str(value),
         }
     }
 }
 
-/// This is a temporary rowid only valid during a transaction. Not persisted, no hash.
-/// It does not keep a txn id around because that is only valid during a transaction.
-/// The in memory representation is always just a TempRowId because the hash for
-/// the rowids in the same transaction will just be the same.
+/// A temporary row ID that is valid only within a transaction.
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub(crate) struct TempRowId(pub(crate) u32);
 
@@ -212,7 +204,7 @@ impl TempRowId {
         }
     }
 
-    pub fn counter(self) -> u32 {
+    pub(crate) fn counter(self) -> u32 {
         self.0
     }
 }
@@ -223,34 +215,7 @@ impl From<u32> for TempRowId {
     }
 }
 
-/// The temporary ops in flight for a transaction
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) enum PendingOp {
-    Add {
-        row_id: TempRowId,
-        table: ir::Path,
-        values: Vec<TxnCellValue>,
-    },
-}
-
-impl PendingOp {
-    pub(crate) fn resolve(&self, commit: CommitHash) -> Op {
-        match self {
-            PendingOp::Add {
-                row_id,
-                table,
-                values,
-            } => Op::Add {
-                row_id: row_id.resolve(commit),
-                table: table.clone(),
-                values: values.iter().map(|value| value.resolve(commit)).collect(),
-            },
-        }
-    }
-}
-
-/// A RowRef is either an existing rowid, or a pending id that belongs to a
-/// a particular transaction.
+/// A reference to an existing row or a pending row in the current transaction.
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub(crate) enum RowRef {
     Existing(RowId),
@@ -278,7 +243,7 @@ impl From<TempRowId> for RowRef {
     }
 }
 
-/// This is the internal representation which is derived from `TxnValue`
+/// The internal transaction representation derived from `TxnValue`.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) enum TxnCellValue {
     Id(RowRef),
@@ -342,35 +307,28 @@ impl From<CellValue> for TxnCellValue {
     }
 }
 
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub enum Op {
+/// An operation staged within a transaction.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub(crate) enum PendingOp {
     Add {
-        row_id: RowId,
-        table: ir::Path, // using path so it's stable across replicas
-        values: Vec<CellValue>,
-    },
-    // TODO Delete + Update
-}
-
-/// Packed representation of an operation.
-#[derive(Debug)]
-pub(crate) enum PackedOp {
-    Add {
-        row_id: PackedRowId,
-        values: Vec<PackedCell>,
+        row_id: TempRowId,
+        table: ir::Path,
+        values: Vec<TxnCellValue>,
     },
 }
 
-impl Op {
-    pub fn id(&self) -> RowId {
+impl PendingOp {
+    pub(crate) fn resolve(&self, commit: CommitHash) -> Op {
         match self {
-            Op::Add { row_id, .. } => *row_id,
-        }
-    }
-
-    pub fn table(&self) -> &ir::Path {
-        match self {
-            Op::Add { table, .. } => table,
+            PendingOp::Add {
+                row_id,
+                table,
+                values,
+            } => Op::Add {
+                row_id: row_id.resolve(commit),
+                table: table.clone(),
+                values: values.iter().map(|value| value.resolve(commit)).collect(),
+            },
         }
     }
 }
